Moved new static pages into catalogue app to minimise clutter in the main folder, updated app urls and tested.
