--- about/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class AboutConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'about'
