--- about/urls.py
@@ -1,10 +1,0 @@
-from django.urls import path
-
-from . import views
-
-app_name = 'about'
-
-urlpatterns = [
-    path('',        views.index,   name='index'),   # /about/
-    path('contact', views.contact, name='contact'), # /about/contact
-]
