--- about/views.py
@@ -1,8 +1,0 @@
-from django.shortcuts import render
-from inertia import render
-
-def index(request):
-    return render(request, 'About/Index')
-
-def contact(request):
-    return render(request, 'About/Contact')
