+++ app/urls.py
@@ -24,7 +24,5 @@ urlpatterns = [
     path('catalogue/', include('catalogue.urls'), name='catalogue'),
     path('livestreams/', include('livestreams.urls'), name='livestreams'),
     path('staff/', include('staff.urls'), name='staff'),
-    path('about/', include('about.urls'), name='about'),
-    path('getinvolved/', include('getinvolved.urls'), name='getinvolved'),
     path('admin/', admin.site.urls),
 ]
